requirements.txt
streamlit
python-dotenv
PyPDF2
spacy
openai
scikit-learn